Simplified mapping
At request of @arwhyte, I've expanded the mapping.  Definitions now refer to only one spec each.  A jq query of this structure is too cumbersome, so I'm switching to lodash for in-program selection of term definitions from mapping.  This part didn't need to use jq anyway.  I'll keep jq (or possibly JSONPath) for the selection of data from the input and manipulating the output.