Switching jq to Lodash, etc.
Changing from jq to Lodash for selecting applicable specs from the mapping.  This isn't complete yet.  The jq version had worked, but was not optimal.  Although I thought that using jq for everything (internally and in the spec mappings) would be nice, it's not necessary.  So I've switched to Lodash for the pieces that don't come from the mapping.  (As I've seen from an xAPI project, I may re-reconsider JSONPath for that instead of jq anyway.)  So far, I was just getting to know Lodash and it's not using the input yet.

I've also homogenized the mapping JSON to always use an array for `domain` and expanded the README a bit.